Editor: Move HTML-mode dropzone higher in DOM to cover editor body
****** TODO: refine this message before merging *****

html mode dropzone introduced in #11680. at the time, there weren't any other dropzone -- the sidebar featured image zone was added a week later, in #11839 -- so the html mode dropzone took up the entire screen.

because it could take up the entire screen, its placement inside the editor-html-toolbar didn't pose any tangible problems.

when a third dropzone was added in #16819, this dropzone was made not-full-screen so that the new dropzone would be accessible. by accident, though, the changes in this PR were not done at that time, which led to the media dropzone being only 10px tall, instead of covering the entire editor body

because there are now multiple dropzones on the editor screen, the html mode dropzone should only cover the editor body itself, not the entire screen. in order to achieve that, it needs to be moved above the editor-html-toolbar, since the toolbar is restricted to 40px tall.